refactor: move primitives types into separate file
